refactor fastai artifacts
